[#41997] new mockup data / import fixes

See merge request gronlandsprojekter/kas!35

+++ kas/worker/templatetags/kas_tags.py
@@ -6,11 +6,11 @@ from django.utils.translation import ngettext as __
 def duration(td):
 
     total_seconds = int(td.total_seconds())
-    days = total_seconds / 86400
+    days = int(total_seconds / 86400)
     remaining_hours = total_seconds % 86400
     remaining_minutes = remaining_hours % 3600
-    hours = remaining_hours / 3600
-    minutes = remaining_minutes / 60
+    hours = int(remaining_hours / 3600)
+    minutes = int(remaining_minutes / 60)
     seconds = (remaining_minutes % 60) + (td.total_seconds() - total_seconds)
 
     return ', '.join([
